'/' to login feature user: Bobby123 user: Sally_55 user: ma_name_jeff All with passwords of 1234
'/feed' to see all tweets

Once logged in, tweets posted from /feed or the /profile/user (e.g /profile/ma_name_jeff) will be directed to the profile page, and posted there with userhandle at top.

Return to log in page and log in again to change the cookies to post on the correct profile page.

Port running on 3001